[PATRO23-37] Rename owned entities
